fire-ignition-detector/requirements.txt
streamlit==1.29.0
torch==2.0.1
torchvision==0.15.2
opencv-python
numpy
Pillow
